requirements.txt
Django==1.6.5
gunicorn==19.0.0
South==1.0
djangorestframework==2.3.14
django-filter==0.7
django-debug-toolbar==1.2.1
django-extensions==1.3.10
Werkzeug==0.9.6
